Add integration test setup
* Add integration test setup

Introduced a `TestKernel` to register the HttpCacheBundle and enabled configuration for FosHttpCache in test environments. Also added a basic integration test to verify the kernel boots properly, ensuring the testing framework is operational. Updated `composer.json` to autoload the test kernel.

* Add possibility to execute the console

* Allow installing Pimcore 10

* Fix lowest deps

* Allow risky tests to fail on lowest deps

Moved the fail-on-risky option to high-dependency test configurations in the CI workflow. Removed its enforcement from the PHPUnit configuration to allow more flexibility for other test setups.